Generated database queries to use for testing automatic query generation.
example-query-gen.js will generate different test queries, along with
the parameters used to create them. This data is stored in
test-queries.js, and will be used to test modules that automatically
assemble queries.

test-queries.js has also been examined to ensure that its output is
correct, in order to prevent the possibility of a query generation bug
being ignored because it also exists in the test query generator.